bpf: tests: condense LXC HostPort test
Run the IPv4 and IPv6 test from a single file.
